Enforce return value units in tests (#61)
Adds enforcement of return value units to the `enforce_units` decorator.
This feature is disabled by default for production and enabled when unit
tests run. The intent is to automate checking of return value units so
any mismatch with the return type annotation is caught during
development. Single return values and tuples of annotated quantities are
both supported.